requirements.txt
opencv-python
numpy
tensorflow
spotipy
Pillow